requirements.txt
# --- Core Framework & API ---
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0

# --- LangChain Ecosystem (Modern Versions) ---
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
langsmith>=0.1.0

# --- Local Embeddings & Vector DB ---
sentence-transformers>=2.2.2
qdrant-client>=1.7.1
unstructured[md]>=0.12.0

# --- Frontend ---
streamlit>=1.30.0